+++ src/TitleDetails/ExpectedPiecesList/ExpectedPiecesList.js
@@ -4,19 +4,7 @@ import PropTypes from 'prop-types';
 import { PIECE_COLUMNS } from '../constants';
 import PiecesList from '../PiecesList';
 
-const visibleColumns = [
-  PIECE_COLUMNS.caption,
-  PIECE_COLUMNS.copyNumber,
-  PIECE_COLUMNS.enumeration,
-  PIECE_COLUMNS.chronology,
-  PIECE_COLUMNS.comment,
-  'format',
-  PIECE_COLUMNS.receiptDate,
-  'request',
-  'selection',
-];
-
-const ExpectedPiecesList = ({ pieces, selectPiece }) => {
+const ExpectedPiecesList = ({ pieces, selectPiece, visibleColumns }) => {
   return (
     <PiecesList
       pieces={pieces}
@@ -30,6 +18,7 @@ const ExpectedPiecesList = ({ pieces, selectPiece }) => {
 ExpectedPiecesList.propTypes = {
   selectPiece: PropTypes.func.isRequired,
   pieces: PropTypes.arrayOf(PropTypes.object),
+  visibleColumns: PropTypes.arrayOf(PropTypes.string).isRequired,
 };
 
 ExpectedPiecesList.defaultProps = {